Moved kfusion and bfusion into include dir and fixed some header includes
