+++ se_denseslam/include/se/bfusion/rendering_impl.hpp
