+++ se_denseslam/include/se/kfusion/rendering_impl.hpp
