+++ se_denseslam/src/rendering.cpp
@@ -6,7 +6,7 @@
  This code is licensed under the MIT License.
 
 
- Copyright 2016 Emanuele Vespa, Imperial College London 
+ Copyright 2016 Emanuele Vespa, Imperial College London
 
  Redistribution and use in source and binary forms, with or without
  modification, are permitted provided that the following conditions are met:
@@ -31,7 +31,7 @@
  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. 
+ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 #include <se/utils/math_utils.h>
 #include <se/commons.h>
@@ -43,14 +43,14 @@
 #include <se/image/image.hpp>
 #include <se/ray_iterator.hpp>
 
-/* Raycasting implementations */ 
-#include "bfusion/rendering_impl.hpp"
-#include "kfusion/rendering_impl.hpp"
+/* Raycasting implementations */
+#include <se/bfusion/rendering_impl.hpp>
+#include <se/kfusion/rendering_impl.hpp>
 
 template<typename T>
 void raycastKernel(const Volume<T>& volume, se::Image<Eigen::Vector3f>& vertex,
    se::Image<Eigen::Vector3f>& normal,
-   const Eigen::Matrix4f& view, const float nearPlane, const float farPlane, 
+   const Eigen::Matrix4f& view, const float nearPlane, const float farPlane,
    const float mu, const float step, const float largestep) {
   TICK();
   int y;
@@ -60,24 +60,24 @@ void raycastKernel(const Volume<T>& volume, se::Image<Eigen::Vector3f>& vertex,
     for (int x = 0; x < vertex.width(); x++) {
 
       Eigen::Vector2i pos(x, y);
-      const Eigen::Vector3f dir = 
+      const Eigen::Vector3f dir =
         (view.topLeftCorner<3, 3>() * Eigen::Vector3f(x, y, 1.f)).normalized();
       const Eigen::Vector3f transl = view.topRightCorner<3, 1>();
       se::ray_iterator<T> ray(*volume._map_index, transl, dir, nearPlane, farPlane);
       ray.next();
       const float t_min = ray.tcmin(); /* Get distance to the first intersected block */
-      const Eigen::Vector4f hit = t_min > 0.f ? 
-        raycast(volume, transl, dir, t_min, ray.tmax(), mu, step, largestep) : 
+      const Eigen::Vector4f hit = t_min > 0.f ?
+        raycast(volume, transl, dir, t_min, ray.tmax(), mu, step, largestep) :
         Eigen::Vector4f::Constant(0.f);
       if(hit.w() > 0.0) {
         vertex[x + y * vertex.width()] = hit.head<3>();
-        Eigen::Vector3f surfNorm = volume.grad(hit.head<3>(), 
+        Eigen::Vector3f surfNorm = volume.grad(hit.head<3>(),
             [](const auto& val){ return val.x; });
         if (surfNorm.norm() == 0) {
           //normal[pos] = normalize(surfNorm); // APN added
           normal[pos.x() + pos.y() * normal.width()] = Eigen::Vector3f(INVALID, 0, 0);
         } else {
-          // Invert normals if SDF 
+          // Invert normals if SDF
           normal[pos.x() + pos.y() * normal.width()] = std::is_same<T, SDF>::value ?
             (-1.f * surfNorm).normalized() : surfNorm.normalized();
         }
@@ -108,8 +108,8 @@ void raycastKernel(const Volume<T>& volume, se::Image<Eigen::Vector3f>& vertex,
 // 	TOCK("renderNormalKernel", normalSize.x * normalSize.y);
 // }
 
-void renderDepthKernel(unsigned char* out, float * depth, 
-    const Eigen::Vector2i& depthSize, const float nearPlane, 
+void renderDepthKernel(unsigned char* out, float * depth,
+    const Eigen::Vector2i& depthSize, const float nearPlane,
     const float farPlane) {
 	TICK();
 
@@ -151,8 +151,8 @@ void renderDepthKernel(unsigned char* out, float * depth,
 	TOCK("renderDepthKernel", depthSize.x * depthSize.y);
 }
 
-void renderTrackKernel(unsigned char* out, 
-    const TrackData* data, 
+void renderTrackKernel(unsigned char* out,
+    const TrackData* data,
     const Eigen::Vector2i& outSize) {
   TICK();
 
@@ -212,19 +212,19 @@ void renderTrackKernel(unsigned char* out,
 }
 
 template <typename T>
-void renderVolumeKernel(const Volume<T>& volume, 
+void renderVolumeKernel(const Volume<T>& volume,
     unsigned char* out, // RGBW packed
-    const Eigen::Vector2i& depthSize, 
-    const Eigen::Matrix4f& view, 
-    const float nearPlane, 
-    const float farPlane, 
+    const Eigen::Vector2i& depthSize,
+    const Eigen::Matrix4f& view,
+    const float nearPlane,
+    const float farPlane,
     const float mu,
-		const float step, 
-    const float largestep, 
+		const float step,
+    const float largestep,
     const Eigen::Vector3f& light,
-		const Eigen::Vector3f& ambient, 
-    bool render, 
-    const se::Image<Eigen::Vector3f>& vertex, 
+		const Eigen::Vector3f& ambient,
+    bool render,
+    const se::Image<Eigen::Vector3f>& vertex,
     const se::Image<Eigen::Vector3f>& normal) {
   TICK();
   int y;
@@ -236,21 +236,21 @@ void renderVolumeKernel(const Volume<T>& volume,
       const int idx = (x + depthSize.x()*y) * 4;
 
       if(render) {
-        const Eigen::Vector3f dir = 
+        const Eigen::Vector3f dir =
           (view.topLeftCorner<3, 3>() * Eigen::Vector3f(x, y, 1.f)).normalized();
         const Eigen::Vector3f transl = view.topRightCorner<3, 1>();
-        se::ray_iterator<typename Volume<T>::field_type> ray(*volume._map_index, 
+        se::ray_iterator<typename Volume<T>::field_type> ray(*volume._map_index,
             transl, dir, nearPlane, farPlane);
         ray.next();
         const float t_min = ray.tmin(); /* Get distance to the first intersected block */
-        hit = t_min > 0.f ? 
-          raycast(volume, transl, dir, t_min, ray.tmax(), mu, step, largestep) : 
+        hit = t_min > 0.f ?
+          raycast(volume, transl, dir, t_min, ray.tmax(), mu, step, largestep) :
           Eigen::Vector4f::Constant(0.f);
         if (hit.w() > 0) {
           test = hit.head<3>();
           surfNorm = volume.grad(test, [](const auto& val){ return val.x; });
 
-          // Invert normals if SDF 
+          // Invert normals if SDF
           surfNorm = std::is_same<T, SDF>::value ? -1.f * surfNorm : surfNorm;
         } else {
           surfNorm = Eigen::Vector3f(INVALID, 0, 0);
